firmata: fix handling of attempted device connect
Previously, the firmata_new() function would call serial_new()
without checking the return value. As a result, the program would
endlessly poll the serial attempting to read bytes it would never
recieve.

This changset frees memory allocated, and returns to the firmata_new()
function with a NULL pointer. The firmat_new() function returns NULL to
the user, and the connection attempt can be aborted without side-effects.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>